+++ obj/src/config/ConfigReader.js
@@ -1,6 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 var _ = require('lodash');
+var mustache = require('mustache');
 var ConfigParams_1 = require("./ConfigParams");
 var ConfigReader = /** @class */ (function () {
     function ConfigReader() {
@@ -12,11 +13,12 @@ var ConfigReader = /** @class */ (function () {
             this._parameters = parameters;
     };
     ConfigReader.prototype.parameterize = function (config, parameters) {
-        // Convert template to lodash
-        config = config.replace(/{{/g, "<%=").replace(/}}/g, "%>");
         parameters = this._parameters.override(parameters);
-        var template = _.template(config);
-        return template(parameters);
+        // Convert template to lodash
+        //config = config.replace(/{{/g, "<%=").replace(/}}/g, "%>");
+        //let template = _.template(config);
+        //return template(parameters);
+        return mustache.render(config, parameters);
     };
     return ConfigReader;
 }());